Testing for /stats endpoint
The first version of the traffic analyzer is complete.
It has the js script to retrieve information about visitors.
An endpoint for sending and storing those values in a MySQL database( one table at the moment)
An endpoint for displaying the data from the database between a time frame.
Testing for each endpoint