Tracking Dashboard
A simple dashboard made with html, css and js.
The dashboard contains 3 parts, a table with users information, a line chart with the number of visits, and a bar chart with number of visits per each page.
I did not went for other technologies since the requirement was not about this.